refactor: Namespace sub-divided into Interfaces and Design patterns
