+++ GoF/Observer.cs
@@ -1,22 +1,6 @@
-namespace ExamPrep
+using ExamPrep.Interfaces;
+namespace ExamPrep.DesignPatterns
 {
-    /// <summary>
-    /// Represents an observer that watches the subject and gets notified about changes.
-    /// </summary>
-    public interface IObserver
-    {
-        /// <summary>
-        /// Gets the name of the observer.
-        /// </summary>
-        /// <returns>The name of the observer.</returns>
-        public string GetName();
-
-        /// <summary>
-        /// Updates the status of the observer.
-        /// </summary>
-        public void UpdateStatus();
-    }
-
     /// <summary>
     /// Represents an observer implementation that subscribes to a subject.
     /// </summary>
@@ -35,32 +19,6 @@ namespace ExamPrep
         {
             Console.WriteLine("\nInside {0}: ObserversCount in Subject = {1}.", _name, _subject.ObserversCount);
         }
-    }
-
-    /// <summary>
-    /// Represents a subject being observed.
-    /// </summary>
-    public interface ISubject
-    {
-        /// <summary>
-        /// Gets or sets the number of observers attached to the subject.
-        /// </summary>
-        public int ObserversCount { get; set; }
-
-        /// <summary>
-        /// Attaches an observer to the subject.
-        /// </summary>
-        public void Attach(IObserver observer);
-
-        /// <summary>
-        /// Detaches an observer from the subject.
-        /// </summary>
-        public void Detach(IObserver observer);
-
-        /// <summary>
-        /// Notifies all attached observers about changes in the subject.
-        /// </summary>
-        public void Notify();
     }
 
     /// <summary>
